fix layer selection in measurement
